unit tests are updated

+++ src/main/java/com/tasks/app/db/TaskService.java
@@ -3,6 +3,10 @@ package com.tasks.app.db;
 import com.google.inject.Inject;
 import com.tasks.app.cache.CacheManager;
 import com.tasks.app.entity.Task;
+
+import javax.swing.text.html.Option;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Response;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -25,13 +29,14 @@ public class TaskService {
         return cacheManager.getTaskFromCache(id);
     }
 
-    public void updateTask(Task task, String id) {
+    public Task updateTask(Task task, String id) {
         taskDAO.updateTask(task, id);
         task.setId(id);
         cacheManager.setTaskToCache(task);
+        return task;
     }
 
-    public void insertTask(Task task) {
+    public Task insertTask(Task task) {
         if (task.getId() == null) {
             String id = UUID.randomUUID().toString();
             task.setId(id);
@@ -40,10 +45,17 @@ public class TaskService {
         } else {
             updateTask(task, task.getId());
         }
+        return task;
     }
 
-    public void deleteTask(String id) {
-        taskDAO.deleteTask(id);
-        cacheManager.deleteTaskFromCache(id);
+    public boolean deleteTask(String id) {
+         if(taskDAO.findTaskById(id).isPresent()) {
+             taskDAO.deleteTask(id);
+             cacheManager.deleteTaskFromCache(id);
+             return true;
+         }
+         else {
+             throw new WebApplicationException("Task not found", 404);
+         }
     }
 }
